Add Visual Studio 2019 extension
This allows supporting the new ELF debugger if used with the 2019
preview.

A separate nupkg (open-enclave-cross) is used that includes the VS
targets files needed by this extension.

Includes support for compiling static libs and C++ files in enclaves

clang is used to compile enclaves and Linux host apps, while the
VS compiler is used for Windows host apps

x64 is supported, but x86 is not

Signed-off-by: Dave Thaler <dthaler@microsoft.com>